Cast to int instead of float because ceil op
Cast to int instead of float because of ceil op. Messages counter will always be an integer.

+++ SMSCounter.php
@@ -151,7 +151,7 @@ class SMSCounter
                 break;
         }
 
-        $messages = ceil($length / $perMessage);
+        $messages = (int) ceil($length / $perMessage);
         $remaining = ($perMessage * $messages) - $length;
 
         $returnset = new \stdClass();
